Removing not needed properties of the copied code
